new serialization switch thread samples
