--- dubbo-demo-switch-thread/src/main/java/com/alibaba/dubbo/samples/DemoService.java
@@ -1,7 +1,0 @@
-package com.alibaba.dubbo.samples;
-
-public interface DemoService {
-
-    String sayHello(String name);
-
-}
