--- dubbo-demo-switch-thread/src/main/java/com/alibaba/dubbo/samples/DemoServiceImpl.java
@@ -1,16 +1,0 @@
-package com.alibaba.dubbo.samples;
-
-import com.alibaba.dubbo.rpc.RpcContext;
-
-import java.text.SimpleDateFormat;
-import java.util.Date;
-
-
-
-public class DemoServiceImpl implements DemoService {
-    @Override
-    public String sayHello(String name) {
-        System.out.println("[" + new SimpleDateFormat("HH:mm:ss").format(new Date()) + "] Hello " + name + ", request from consumer: " + RpcContext.getContext().getRemoteAddress());
-        return "Hello " + name + ", response from provider: " + RpcContext.getContext().getLocalAddress();
-    }
-}
