--- dubbo-demo-switch-thread/src/main/resources/dubbo.properties
@@ -1,1 +1,0 @@
-dubbo.application.qos.port=22222
