+++ pom.xml
@@ -55,6 +55,7 @@
         <module>dubbo-samples-zookeeper</module>
         <module>dubbo-samples-multi-registry</module>
         <module>dubbo-demo-switch-thread</module>
+        <module>dubbo-samples-switch-serialization-thread</module>
     </modules>
 
     <properties>
